feat: error re-exports
