+++ crates/derive/src/errors/mod.rs
@@ -1,0 +1,15 @@
+//! Error types for kona's derivation pipeline.
+
+mod attributes;
+pub use attributes::BuilderError;
+
+mod stages;
+pub use stages::BatchDecompressionError;
+
+mod pipeline;
+pub use pipeline::{
+    PipelineEncodingError, PipelineError, PipelineErrorKind, PipelineResult, ResetError,
+};
+
+mod sources;
+pub use sources::{BlobDecodingError, BlobProviderError};
