+++ crates/derive/src/pipeline/mod.rs
@@ -1,14 +1,5 @@
 //! Module containing the derivation pipeline.
 
-/// Re-export trait arguments.
-pub use crate::traits::{
-    AttributesBuilder, DataAvailabilityProvider, NextAttributes, OriginAdvancer, OriginProvider,
-    Pipeline, ResetProvider, Signal, SignalReceiver, StepResult,
-};
-
-/// Re-export error types.
-pub use crate::errors::{PipelineError, PipelineResult};
-
 mod builder;
 pub use builder::PipelineBuilder;
