+++ crates/derive/src/stages/batch/mod.rs
@@ -11,7 +11,7 @@
 //! [ChannelReader]: crate::stages::channel::ChannelReader
 //! [AttributesQueue]: crate::stages::attributes_queue::AttributesQueue
 
-use crate::pipeline::PipelineResult;
+use crate::errors::PipelineResult;
 use alloc::boxed::Box;
 use async_trait::async_trait;
 use op_alloy_protocol::{Batch, BlockInfo, L2BlockInfo};
